% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ecdf.R
\name{create_ecdf}
\alias{create_ecdf}
\title{Create Empirical Cumulative Distribution Function (ECDF)}
\usage{
create_ecdf(
  vector,
  title = "ECDF Chart",
  subtitle = "",
  caption = "",
  x_label = "Quartiles",
  y_label = "Frequency"
)
}
\arguments{
\item{vector}{Vector containing the observed numeric data.}

\item{title}{Character string; title to be displayed on Pareto chart.  Default
is "ECDF Plot".}

\item{subtitle}{Character string; subtitle to be displayed on chart.  Default
is "".}

\item{caption}{Character string; caption to be displayed on the chart.  Default
is "".}

\item{x_label}{Character string; x axis label to be displayed on the chart.
Default is "Groups".}

\item{y_label}{Character string; y axis label to be displayed on the chart.
Default is "Frequency".}
}
\value{
A plot containing the ECDF.  If an error or warning occurs,
a message will be printed to the console and the function will return NULL.
}
\description{
\code{create_ecdf} generates a plot of the Empirical Cumulative Distribution
Function (ECDF) using the observed data.  The ECDF shows the proportion of
obervations that are less than or equal to each obvsered value.
}
\examples{
source <- c(1, 1, 1, 1, 2, 2, 2, 3, 3, 4)
plot <- create_ecdf(source)

}
