% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_groups.R
\name{create_groups}
\alias{create_groups}
\title{Create Groups}
\usage{
create_groups(source, low, high, step)
}
\arguments{
\item{source}{Vector, list or data frame input.  If the source is of class
list, it will be simplified to a vector using \code{unlist}.}

\item{low}{Integer, the lowest value used to generate the groups.}

\item{high}{Integer, the highest value used to generate the groups.}

\item{step}{Integer, the interval used to break up the range of
discrete values.}
}
\value{
A data frame containing the group values.  If an error or warning
occurs, a message will be printed to the console and the function will
return NULL.
}
\description{
\code{create_groups} groups discrete values into buckets that can be used to
generate a frequency distribution table.  This function is useful when the
data has so many values it would not make sense to generate the table with
the values themselves.
}
\details{
\code{create_groups} replaces each discrete value with the
appropriate group based on the parameters passed to the function.
}
\examples{
\dontrun{
table <- c(10, 20, 25, 30, 35, 35,
           35, 40, 50, 75, 85, 95)
table <- create_groups(table, 10, 100, 10)
}

}
