% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get Groups}
\usage{
get_groups(observations)
}
\arguments{
\item{observations}{An integer representing the number of observations to be
grouped.}
}
\value{
An integer representing the result of the logarithm calculation.
If an error or warning occurs, a message will be printed to the console and
the function will return NULL.
}
\description{
\code{get_groups} calculates a number of groups for a number of observations
using the logarithm calculation \code{log(observations) / log(2)}.
}
\details{
\code{get_groups} expects to receive an integer as an input value.
If the function is passed a non-integer value, e.g. 9.5, it will round the
value up to the nearest integer.
}
\examples{
\dontrun{
number_of_groups <- get_groups(1000)
}

}
